==> 3_feature_generation.ipynb <==
### Генерация признаков и сохранения их

генерируем 3 набора - feature_pack_v1,feature_pack_v2,feature_pack_v3

- feature_pack_v1 - основные признаки выделенные из данных
- feature_pack_v2 - основные признаки выделенные из данных и их преобразования
- feature_pack_v3 - основные признаки выделенные из данных + лаговые агрегаты


import os
if 'drive' not in os.listdir():
    from google.colab import drive
    drive.mount('/content/drive')

path_to_folder = '/content/drive/MyDrive/psb_hack'

# standard libraries
import sys
import os
import json
import re
from tqdm import tqdm
import warnings
warnings.filterwarnings('ignore')
# ds libraries
import pandas as pd
import numpy as np
import joblib
import gc
# свои функции по оценке метрик
path_to_model_utils = '/content/drive/MyDrive/psb_hack'
sys.path.append(os.path.join(path_to_model_utils,'model_utils'))
from features_pack import *

## Загрузка данных

# чтение данных
dates = ['Дата бронирования','Дата отмены', 'Заезд', 'Выезд']
train_full = pd.read_csv(os.path.join(path_to_folder,'data','full_train.csv'),parse_dates=dates)
train = pd.read_csv(os.path.join(path_to_folder,'data','train.csv'),parse_dates=dates)
oot_val = pd.read_csv(os.path.join(path_to_folder,'data','oot_val.csv'),parse_dates=dates)
oof_val = pd.read_csv(os.path.join(path_to_folder,'data','oof_val.csv'),parse_dates=dates)
test_data = pd.read_csv(os.path.join(path_to_folder,'data','test.csv'),parse_dates=list(filter(lambda x: x != 'Дата отмены',dates)))

gc.collect()

## Features pack 1

# расчет фичей
train_full_v1 = features_pack_v1(train_full)
train_v1 = features_pack_v1(train)
oot_val_v1 = features_pack_v1(oot_val)
oof_val_v1 = features_pack_v1(oof_val)
test_data_v1 = features_pack_v1(test_data)
# сохранение фичей
# сохраним все в csv, чтобы быстрее было считывать
train_full_v1.to_csv(os.path.join(path_to_folder,'data','full_train_features_pack_v1.csv'),header = True,index=False)
train_v1.to_csv(os.path.join(path_to_folder,'data','train_features_pack_v1.csv'),header = True,index=False)
oot_val_v1.to_csv(os.path.join(path_to_folder,'data','oot_val_features_pack_v1.csv'),header = True,index=False)
oof_val_v1.to_csv(os.path.join(path_to_folder,'data','oof_val_features_pack_v1.csv'),header = True,index=False)
test_data_v1.to_csv(os.path.join(path_to_folder,'data','test_features_pack_v1.csv'),header = True,index=False)

## Features pack 2

# расчет фичей
train_full_v2 = features_pack_v2(train_full)
train_v2 = features_pack_v2(train)
oot_val_v2 = features_pack_v2(oot_val)
oof_val_v2 = features_pack_v2(oof_val)
test_data_v2 = features_pack_v2(test_data)
# сохранение фичей
# сохраним все в csv, чтобы быстрее было считывать
train_full_v2.to_csv(os.path.join(path_to_folder,'data','full_train_features_pack_v2.csv'),header = True,index=False)
train_v2.to_csv(os.path.join(path_to_folder,'data','train_features_pack_v2.csv'),header = True,index=False)
oot_val_v2.to_csv(os.path.join(path_to_folder,'data','oot_val_features_pack_v2.csv'),header = True,index=False)
oof_val_v2.to_csv(os.path.join(path_to_folder,'data','oof_val_features_pack_v2.csv'),header = True,index=False)
test_data_v2.to_csv(os.path.join(path_to_folder,'data','test_features_pack_v2.csv'),header = True,index=False)

## Features pack 3

# расчет фичей
train_full_v3 = features_pack_v3(train_full,'full')
train_v3 = features_pack_v3(train,'sample')
oot_val_v3 = features_pack_v3(oot_val,'sample')
oof_val_v3 = features_pack_v3(oof_val,'sample')
test_data_v3 = features_pack_v3(test_data,'full')
# сохранение фичей
# сохраним все в csv, чтобы быстрее было считывать
train_full_v3.to_csv(os.path.join(path_to_folder,'data','full_train_features_pack_v3.csv'),header = True,index=False)
train_v3.to_csv(os.path.join(path_to_folder,'data','train_features_pack_v3.csv'),header = True,index=False)
oot_val_v3.to_csv(os.path.join(path_to_folder,'data','oot_val_features_pack_v3.csv'),header = True,index=False)
oof_val_v3.to_csv(os.path.join(path_to_folder,'data','oof_val_features_pack_v3.csv'),header = True,index=False)
test_data_v3.to_csv(os.path.join(path_to_folder,'data','test_features_pack_v3.csv'),header = True,index=False)

## Features pack 4

# расчет фичей
train_full_v4 = features_pack_v4(train_full)
train_v4 = features_pack_v4(train)
oot_val_v4 = features_pack_v4(oot_val)
oof_val_v4 = features_pack_v4(oof_val)
test_data_v4 = features_pack_v4(test_data)
# сохранение фичей
# сохраним все в csv, чтобы быстрее было считывать
train_full_v4.to_csv(os.path.join(path_to_folder,'data','full_train_features_pack_v4.csv'),header = True,index=False)
train_v4.to_csv(os.path.join(path_to_folder,'data','train_features_pack_v4.csv'),header = True,index=False)
oot_val_v4.to_csv(os.path.join(path_to_folder,'data','oot_val_features_pack_v4.csv'),header = True,index=False)
oof_val_v4.to_csv(os.path.join(path_to_folder,'data','oof_val_features_pack_v4.csv'),header = True,index=False)
test_data_v4.to_csv(os.path.join(path_to_folder,'data','test_features_pack_v4.csv'),header = True,index=False)
