==> 7_choose_best_config_model.ipynb <==
# Выбор оптимальной конфигурации модели

# standard libraries
import pandas as pd
import sys
import os
import yaml
import warnings

warnings.filterwarnings('ignore')
pd.set_option('display.float_format','{:_.2f}'.format)
if 'drive' not in os.listdir():
    from google.colab import drive
    drive.mount('/content/drive')

path_to_folder = '/content/drive/MyDrive/psb_hack'

# Выбор лучшей модели по метрикам по продукту

main_metrics = ['precision', 'recall', 'f1', 'roc-auc original', 'roc-auc_hack']
test_set_name = 'val_oot'

def read_metric_table(file_name):
    table = pd.read_excel(file_name)
    table['config_model'] = file_name.split('/')[-1].split('.')[0].replace('_metrics','')
    return table

# Загрзука метрик

# файлы с метриками по модели суммы кредита
metrics = os.listdir(os.path.join(path_to_folder, 'metrics'))
# соберем их в единый файл с указанием модели
all_metrics = pd.concat(list(map(lambda x: read_metric_table(os.path.join(path_to_folder,'metrics', x)),metrics))).reset_index(drop=True)

all_metrics.sample(5)

### Выбор лучшей конфигурации на oot

print('Лучашя модель по метрика')
display(all_metrics.loc[(all_metrics['dataset']=='oot_val') & (all_metrics['group']=='overall')][main_metrics] \
              .applymap(lambda x: all_metrics.iloc[int(x),-1]))
print()
oot = all_metrics.loc[all_metrics['dataset']=='oot_val'] \
              .groupby(['group','measurement'])[main_metrics].idxmin().applymap(lambda x: all_metrics.iloc[int(x),-1])
print('лучшая модель для каждой группы')
display(oot)

### Выбор лучшей конфигурации на val

print('Лучашя модель по метрикам')
display(all_metrics.loc[(all_metrics['dataset']=='oof_val') & (all_metrics['group']=='overall')][main_metrics] \
              .applymap(lambda x: all_metrics.iloc[int(x),-1]))
print()
oof = all_metrics.loc[all_metrics['dataset']=='oof_val'] \
              .groupby(['group','measurement'])[main_metrics].idxmin().applymap(lambda x: all_metrics.iloc[int(x),-1])
print('лучшая модель для каждой группы')
display(oot)
