==> 8_make_prediction.ipynb <==
# Предикт на сабмит


# установка последнего lightautoml
# !pip install -q git+https://github.com/sb-ai-lab/LightAutoML.git

import os
if 'drive' not in os.listdir():
    from google.colab import drive
    drive.mount('/content/drive')

path_to_folder = '/content/drive/MyDrive/psb_hack'

# standard libraries
import sys
import os
import json
import re
from tqdm import tqdm
import warnings
warnings.filterwarnings('ignore')
# ds libraries
import pandas as pd
import numpy as np
# lama
from lightautoml.automl.presets.tabular_presets import TabularAutoML
from lightautoml.tasks import Task
# save/load model
import joblib
import gc
import yaml

### Загрузка конфига и данных

print("Доступные конфиги модели:\n")
sorted(filter(lambda x: 'lama' in x, os.listdir(os.path.join(path_to_folder, 'models'))))

# конфиг модели
config_name = 'psb_hack_lama_v541.yaml'

# конфиг модели
with open(os.path.join(path_to_folder, 'configs', config_name),'r',encoding='utf-8') as config:
    config_model = yaml.full_load(config)
# нужные файлы
full_train_flag = config_model['full_train_flag']
features_pack = config_model['features_pack']
model_file =  os.path.join(path_to_folder,'models',config_model['model_name'] + '.pkl')
metric_file =  os.path.join(path_to_folder,'metrics',config_model['model_name'] + '_metrics.xlsx')

# чтение данных
dates = ['Дата бронирования','Заезд', 'Выезд']
read_cols = None
test = pd.read_csv(os.path.join(path_to_folder,'data',f'test_{features_pack}.csv'),parse_dates=dates,usecols=read_cols)
gc.collect()
# загрузим модель

with open(model_file,'rb') as save_model:
    model = joblib.load(save_model)
print(model.create_model_str_desc())

# загрузка теста и шаблона сэмпла
sample = pd.read_csv(os.path.join(path_to_folder,'data','example.csv'),names = ['pred'])
# предикты и сохранение
sample['pred'] = np.around(model.predict(test).data.reshape(-1),3)#(model.predict(test).data.reshape(-1) >= thr).astype(int)
sample.to_csv(os.path.join(path_to_folder,'submits',config_model['model_name'] + '_sample.csv'),header = False,index=False)
# проверка что все сохранилось
pd.read_csv(os.path.join(path_to_folder,'submits',config_model['model_name'] + '_sample.csv'),names=['pred']).info()
